mcp/requirements/requirements-core.txt
langchain-mcp-adapters==0.0.5
python-dotenv==1.1.0